Step 6: Analyze manual test failures and refactor source code based on results. - Recorded bug reports under the analysis folder. - Logged LLM-assisted fix prompts in llm_logs. - Resolved edge cases and logic errors identified during manual testing.

+++ src/main/java/humaneval/llm1/task31/Solution.java
@@ -31,7 +31,7 @@ public class Solution {
         if (n % 2 == 0 || n % 3 == 0) {
             return false;
         }
-        for (int i = 5; i * i <= n; i = i + 6) {
+        for (int i = 5; (long) i * i <= n; i = i + 6) {
             if (n % i == 0 || n % (i + 2) == 0) {
                 return false;
             }
